+++ packages/ag-grid/src/lib/types/index.ts
@@ -1,2 +1,3 @@
 export * from './gridConfig';
 export * from './gridController';
+export * from './proxy';
